+++ index.html
@@ -15,7 +15,7 @@
 </head>
 <body>
 
-  <div id="app"></div>
+  <div id="root"></div>
   <noscript>
     You need to enable Javascript to run this React application
   </noscript>
