+++ src/MainPage.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div class="ui three column centered grid">
+        <div class="row">
+      <h1>Facundo Osores</h1>
+      </div>
+      <div></div>
+        <div class="row">
+      <h2>Front End Developer</h2>
+      </div>
+        <div class="row">
+      <h4>Creative, revolutionary, passionate</h4>
+      </div>
+
+    </div>
+  )
+}
+
+export default Home
