+++ src/index.js
@@ -1,31 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Hello from "./Hello";
-import Header from "./Header";
-import Footer from "./Footer";
+import App from "./App";
 import { Switch, Route, BrowserRouter } from "react-router-dom";
-import Projects from "./Projects";
-import About from "./About";
-import Cv from "./Cv";
 
-const App = () => {
-  return (
-    <>
-      <Header />
-      <Switch>
-        <Route exact path="/" component={Hello}></Route>
-        <Route exact path="/about" component={About}></Route>
-        <Route exact path="/projects" component={Projects}></Route>
-        <Route exact path="/cv" component={Cv}></Route>
-      </Switch>
-      <Footer />
-    </>
-  );
-};
 
 ReactDOM.render(
   <BrowserRouter>
     <App />
   </BrowserRouter>,
-  document.getElementById("app")
+  document.getElementById("root")
 );
